# 8319184: RISC-V: improve MD5 intrinsic