@gctony 2023-11-01
Performance testing on my VisionFive 2 shows a mild improvement, usually between 1.5%-2%. Here are the jmicro results:

before (with +UseZbb):

```
Benchmark                    (digesterName)  (length)  (provider)  Mode  Cnt       Score    Error  Units
MessageDigests.digest                   md5        64     DEFAULT  avgt  300    1687.068 ?  2.670  ns/op
MessageDigests.digest                   md5     16384     DEFAULT  avgt  300  136355.127 ? 16.621  ns/op
MessageDigests.getAndDigest             md5        64     DEFAULT  avgt  300    2402.758 ?  3.744  ns/op
MessageDigests.getAndDigest             md5     16384     DEFAULT  avgt  300  138048.866 ? 24.224  ns/op
```

after (with +UseZbb):

```
Benchmark                    (digesterName)  (length)  (provider)  Mode  Cnt       Score    Error  Units
MessageDigests.digest                   md5        64     DEFAULT  avgt  300    1678.336 ?  1.056  ns/op
MessageDigests.digest                   md5     16384     DEFAULT  avgt  300  134529.688 ? 14.527  ns/op
MessageDigests.getAndDigest             md5        64     DEFAULT  avgt  300    2384.615 ?  4.325  ns/op
MessageDigests.getAndDigest             md5     16384     DEFAULT  avgt  300  136159.783 ? 23.986  ns/op
```